uploading my triangle reducer script for minecraft based voxel type
using the Craig Perko old Minecraft tutorial found here https://www.youtube.com/watch?v=YpHQ-Kykp_s&ab_channel=CraigPerko
for the chunks normal 2 triangles per cubic face. I changed the logic in order for cubic faces to merge together and be a single larger face, when the triangles are on the same plane and that it forms a cube or a rectangle.

It was long to figure it out and a lot of code is involved. I bet there are better algorithms elsewhere, but i am glad i was able to come up with this one.

thank you for reading me.
steve chassé