uploaded my 2018 script
both versions are the same thing and the 2020 script edited script i only removed the useless variables available in the editor but i left tons of commented useless code for my drafts. So using the settings of the unity3d inspector on the image of the readme file for the fake triangulator i have invented without using references, because i wanted to bang my head on scripting for minutes and hours and days trying to figure this shit out. It's probably the lamest of the lamest of the triangulators of them all, but i love the square and rectangular faces it gives and the highly reduced number of vertices and triangles on a simple minecraft chunk.

I haven't learned RayTracing/RayCasting rendering and how it works yet so i'm gonna still be using directX11 until i can no longer work in it and because of that, i have to stick to "less triangles/less vertices" to render. I know. it's lame but i don't care. why wouldn't i go and use the already provided triangulation techniques and shut the hell up? Well, i've worked a lot for this tiny little script, thinking and thinking and working so why wouldn't i be proud of releasing it if i figured out my own way of triangulating a chunk like i have? but i coded this script in september 2018... By now, theyre are probably even faster triangulators available everywhere... But in 2018, i couldn't find any that i wanted to use although a lot were working but none was using this type of fake triangulation. So i banged my head on walls of scripting errors, because even after having started scripting in 2016-2017, i still was very bad like all will see reading those old scripts. I couldn't achieve marching cubes yet and i wanted to increase the performance of an infinite minuscule chunk minecraft genre game. I thought those marching cubes and marching squares scripts were too hard to understand for me in 2018 and i wouldn't want to try marching cubes yet because of that. I couldn't find a triangulator like the one i wanted to build, so i decided to code it. Also back then i had read that using an underscore before variable names was a great idea... So i coded like that for a while, underscore with ALL of the TOTAL of my variable names. i read that it was a great idea somewhere on programming forums to underscore name variables. i don't script like that anymore.

sometimes, i don't clean my drafts and i don't care releasing my drafts on github. and for a fact, it's just more authentic, until i post all of my scripting
videos of hours and hours on twitch.tv where i was so lonely and without much views. But with tons of mp4 videos on how i code my stuff, if i would edit them and remove all of the crap i say when i script and program, maybe i could start and make a buck on youtube, otherwise without editing my videos they would be banned from youtube for swearing and being a complotists scenario inventor for video games scenarios and conspirationists scenario inventor for video games scenarios... Edited and my junk removed, maybe i would get paid for scripting hours and hours for useless scripts. It's a 2018 edition people. i don't use underscore of variable names anymore and in my next cleanup version, i will try and make this more performant.

thank you for reading me,
steve chassé